% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_converter.R
\name{FIPS_converter}
\alias{FIPS_converter}
\title{Convert coordinates to FIPS}
\usage{
FIPS_converter(id, lat, long, state_data, dataset = NULL)
}
\arguments{
\item{id}{Id to be preserved in the output}

\item{lat}{Latiutude}

\item{long}{Longitude}

\item{state_data}{Shapefiles for the state}

\item{dataset}{The data set you would like to join the results onto}
}
\description{
Convert geocoded coordinates to a FIPS code. Can be used for a data set or a single data point.
}
\examples{
Mayo = grab_shapefiles(c("MN", "WI"))
FIPS_converter(id = c("House 1", "House 2", "House 3"),
               lat = c(44.079216, 45.8283963, 46.1252037),
               long = c(-92.66854, -95.4019813, -95.4569129),
               state_data = Mayo)

            id         FIPS
1 Some house 1 271090019002
2 Some house 2 270419509001
3 Some house 3 271119615001
}
