% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_shapefiles.R
\name{grab_shapefiles}
\alias{grab_shapefiles}
\title{Download TIGER Shapefiles}
\usage{
grab_shapefiles(state_abbrv)
}
\arguments{
\item{state_abbrv}{The standard two-letter state abbreviation}
}
\description{
Downlaod shapefiles for a state based on the state abbreviation.
}
\examples{
MN = grab_shapefiles("MN")
Mayo_shapefiles = grab_shapefiles(c("MN", "WI"))
plot(Mayo_shapefiles) # Be careful, this is slow!
whole_us = grab_shapefiles("everything")
}
